+++ app/javascript/vue/components/radials/ce/components/GeoreferenceSlice.vue
@@ -86,6 +86,7 @@ import UpdateBatch from '@/components/radials/shared/UpdateBatch.vue'
 import VMap from '@/components/ui/VMap/VMap.vue'
 import SmartSelector from '@/components/ui/SmartSelector.vue'
 import SmartSelectorItem from '@/components/ui/SmartSelectorItem.vue'
+import updateMessage from '../utils/updateMessage.js'
 
 const MAX_LIMIT = 250
 
@@ -145,14 +146,6 @@ function reset() {
   collectingEvent.value = null
   georeferences.value = []
   selectedGeoreferences.value = []
-}
-
-function updateMessage(data) {
-  const message = data.sync
-    ? `${data.updated.length} collecting events queued for updating.`
-    : `${data.updated.length} collection events were successfully updated.`
-
-  TW.workbench.alert.create(message, 'notice')
 }
 
 function setCollectingEvent(ce) {