+++ app/javascript/vue/tasks/digitize/store/actions/loadTypeSpecimens.js
@@ -1,21 +1,37 @@
-import { TypeMaterial } from '@/routes/endpoints'
+import { TypeMaterial, Citation } from '@/routes/endpoints'
 import { MutationNames } from '../mutations/mutations'
+import { TYPE_MATERIAL } from '@/constants'
 import makeTypeMaterial from '@/factory/TypeMaterial.js'
+import makeCitation from '@/factory/Citation.js'
 
-export default ({ commit }, id) =>
-  new Promise((resolve, reject) => {
-    TypeMaterial.where({
-      collection_object_id: id,
-      extend: ['roles', 'origin_citation']
-    }).then(
-      ({ body }) => {
-        const typeMaterials = body.map((item) => makeTypeMaterial(item))
+export default async ({ commit }, id) => {
+  try {
+    const { body: typeMaterialData } = await TypeMaterial.where({
+      collection_object_id: id
+    })
 
-        commit(MutationNames.SetTypeMaterials, typeMaterials)
-        resolve(typeMaterials)
-      },
-      (error) => {
-        reject(error)
+    const typeMaterials = typeMaterialData.map(makeTypeMaterial)
+    const typeMaterialIds = typeMaterials.map((item) => item.id)
+
+    const { body: citations } = await Citation.where({
+      citation_object_id: typeMaterialIds,
+      citation_object_type: TYPE_MATERIAL,
+      is_original: true
+    })
+
+    const typeMaerialsWithCitations = typeMaterials.map((item) => {
+      const citation = citations.find((c) => c.citation_object_id === item.id)
+
+      return {
+        ...item,
+        citation: { ...makeCitation(TYPE_MATERIAL), ...citation }
       }
-    )
-  })
+    })
+
+    commit(MutationNames.SetTypeMaterials, typeMaerialsWithCitations)
+
+    return typeMaerialsWithCitations
+  } catch (error) {
+    throw error
+  }
+}