+++ spec/models/asserted_distribution_spec.rb
@@ -595,7 +595,11 @@ describe AssertedDistribution, type: :model, group: [:geo, :shared_geo] do
       # Cause the async delayed job to run immediately here.
       allow_any_instance_of(ActiveSupport::Duration).to receive(:from_now).and_return(Time.now)
 
-      AssertedDistribution.batch_update(params.merge({ async_cutoff: 1 }))
+      AssertedDistribution.batch_update(params.merge({
+        async_cutoff: 1,
+        user_id: Current.user_id,
+        project_id: Current.project_id
+      }))
 
       expect(ad1.reload.asserted_distribution_shape_id).to eq(ga1.id)
 
@@ -610,6 +614,39 @@ describe AssertedDistribution, type: :model, group: [:geo, :shared_geo] do
       expect(r.total_attempted).to eq(2)
       expect(r.updated.count).to eq(2)
       expect(r.not_updated.count).to eq(0)
+    end
+
+    specify 'can set no_dwc_occurrence: true on otu' do
+      ad = AssertedDistribution.create!(asserted_distribution_object: otu,
+        asserted_distribution_shape: geographic_area,
+        source:,
+        no_dwc_occurrence: true
+      )
+
+      expect(ad.no_dwc_occurrence).to be_truthy
+      expect(ad.dwc_occurrence).to be_falsey
+    end
+
+    specify 'default is no_dwc_occurrence: false on otu' do
+      ad = AssertedDistribution.create!(asserted_distribution_object: otu,
+        asserted_distribution_shape: geographic_area,
+        source:
+      )
+
+      expect(ad.no_dwc_occurrence).to be_falsey
+      expect(ad.dwc_occurrence).to be_truthy
+    end
+
+    specify 'no_dwc_occurrence is always true on non-otu' do
+      ba = FactoryBot.create(:valid_biological_association)
+      ad = AssertedDistribution.create!(asserted_distribution_object: ba,
+        asserted_distribution_shape: geographic_area,
+        source:,
+        no_dwc_occurrence: false
+      )
+
+      expect(ad.no_dwc_occurrence).to be_truthy
+      expect(ad.dwc_occurrence).to be_falsey
     end
   end